New mode "direct" to set individual LEDs to individual colors
Pass a JSON object (e.g. using the JSON Template node and passing {{msg}}
in the parameter) containing a key named "groups" which is a list of
objects with "color" (string color name/hascode) and "leds" (a list of
integer LED index with the first LED on the wire being zero) parameters.

{
  "groups":[
    {
      "color":"blue",
      "leds":[1,3,5,9]
    },
    {
      "color":"red",
      "leds":[0,10,132]
    }
  ]
}